+++ group_2/src/table.hpp
@@ -15,8 +15,8 @@ using namespace std;
 
 struct Table
 {
-    Table(int newAlpha);
-    int alpha;
+    Table(double newAlpha);
+    double alpha;
     //unordered_map<size_t, unordered_map<char, uint>> table;
     //google::sparse_hash_map<size_t, google::sparse_hash_map<char, uint>> table;
     ska::flat_hash_map<size_t, ska::flat_hash_map<char, uint>> table;
